Update Grid.java
Grid completly tested :) all correct
